requirements-dev.txt
pylint
pydocstyle
pylint-django
black
